- move the time keyword check_files test  to its own source file. It  only runs on linux now.
